analytics

+++ lib/providers/user_provider.dart
@@ -232,7 +232,7 @@ class UserProvider extends ChangeNotifier {
                 gender: user["gender"],
                 isBlocked: user["blocked"],
                 userType: assignUserType(user["userType"]),
-                rating: user["rating"] ?? 0,
+                rating: (user["rating"] ?? 0).toDouble(),
               ));
             }
           } else if (loggedInUserType == UserType.manager) {
@@ -251,7 +251,7 @@ class UserProvider extends ChangeNotifier {
                 gender: user["gender"],
                 isBlocked: user["blocked"],
                 userType: assignUserType(user["userType"]),
-                rating: user["rating"] ?? 0,
+                rating: (user["rating"] ?? 0).toDouble(),
               ));
             }
           }
